rtl8188eu: Replace LONG variants with the standard uxx form

+++ include/drv_types.h
@@ -165,9 +165,9 @@ struct registry_priv {
 
 
 /* For registry parameters */
-#define RGTRY_OFT(field) ((ULONG)FIELD_OFFSET(struct registry_priv,field))
+#define RGTRY_OFT(field) ((u32)FIELD_OFFSET(struct registry_priv,field))
 #define RGTRY_SZ(field)   sizeof(((struct registry_priv*) 0)->field)
-#define BSSID_OFT(field) ((ULONG)FIELD_OFFSET(struct wlan_bssid_ex,field))
+#define BSSID_OFT(field) ((u32)FIELD_OFFSET(struct wlan_bssid_ex,field))
 #define BSSID_SZ(field)   sizeof(((struct wlan_bssid_ex *) 0)->field)
 
 #define is_primary_adapter(adapter) (1)
